# Conflicts:
#	Assets/Naninovel/Materials/RevealableText.mat
#	Assets/Prefabs/Object Prefabs/FinalCrystal.prefab.meta
#	Assets/Sounds.meta
#	Assets/Texts/Maps/map2-1Mat.txt.meta

+++ Assets/Scripts/Sprites/ObjectAnimator.cs
@@ -6,7 +6,7 @@ public class ObjectAnimator : MonoBehaviour
 {
     protected Renderer sRender;
     public int maxFrames;
-    public int framesPerSecond = 6;
+    public float framesPerSecond = 6;
     protected float timeSinceLastFrame = 0;
     protected float offsetFix = .00001f;
     protected int currentFrame = 0;
